+++ op-challenger/fault/agent.go
@@ -29,8 +29,11 @@ func NewAgent(loader Loader, maxDepth int, trace TraceProvider, responder Respon
 }
 
 // Act iterates the game & performs all of the next actions.
-func (a *Agent) Act() error {
-	game, err := a.newGameFromContracts(context.Background())
+func (a *Agent) Act(ctx context.Context) error {
+	if a.tryResolve(ctx) {
+		return nil
+	}
+	game, err := a.newGameFromContracts(ctx)
 	if err != nil {
 		a.log.Error("Failed to create new game", "err", err)
 		return err
@@ -48,6 +51,19 @@ func (a *Agent) Act() error {
 		}
 	}
 	return nil
+}
+
+// tryResolve resolves the game if it is in a terminal state
+// and returns true if the game resolves successfully.
+func (a *Agent) tryResolve(ctx context.Context) bool {
+	if a.responder.CanResolve(ctx) {
+		err := a.responder.Resolve(ctx)
+		if err != nil {
+			return true
+		}
+		a.log.Error("failed to resolve the game", "err", err)
+	}
+	return false
 }
 
 // newGameFromContracts initializes a new game state from the state in the contract