+++ endpoint-monitor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @eth-optimism/endpoint-monitor
 
+## 1.0.2
+
+### Patch Changes
+
+- [#6164](https://github.com/ethereum-optimism/optimism/pull/6164) [`c11039060`](https://github.com/ethereum-optimism/optimism/commit/c11039060bc037a88916c2cba602687b6d69ad1a) Thanks [@pengin7384](https://github.com/pengin7384)! - fix typo
+
 ## 1.0.1
 
 ### Patch Changes